Renames ToastOverlayScreen to MessageOverlayScreen; Fixes wrong copy on beta access removal.
